[PT-BR] Fix typos and anchors on summaries
Some words for the summaries has special characters like `ç`, `õ`, `ã`
that Jekyll skips when compile the markdown and make the anchors.
It prevented the Summary on pages that have headers with these
characters to work and a custom `<a name=""></a>` was necessary to
make the Summary work again.

I also fixed some typos on the summary that also prevented the summary
to work.