Add arity to the chunk_by description in basics Enum guide (#537)
This is RE @nscyclone 's comment in [this
thread](https://github.com/doomspork/elixir-school/issues/529#issuecomment-233766599)
discussing the specific elaboration on Enum.chunk/1. In other areas it
has been conventional to reference arity and it was not done here. This
clarifies this and removes ambiguity as to whether it could be
Enum.chunk_by/3.