#828 Plug lesson fixes (#861)
* #828 Fixes to Plug lesson

- Include the `mix deps.get` command
- Eliminate one of two `def applications` versions
- Remove import `Plug.Conn` line from `verify_request.ex`

* #828 Plug: Break down lesson into smaller chunks

The lesson starts with a bare minimum Plug application and verifies it. It then goes into `Plug.Router`,
adding plugs to the router, and configuring the HTTP port.

This also tweaks module names to be more in line with Elixir module and filename conventions.

* #828 One line per sentence

Also, remove double spaces after periods.