Update plug.md
I believe elixir 1.4 introduced the change of applications -> extra_applications, where you don't need to add an app to start it, it is assumed from the dependencies. The original English lesson has the code chnaged but not the text around it. Should we remove this text, or let it with an asterisk as a backwards compatible option?