Update enum.md (#1128)
* Update enum.md

Bump version with some minor tweaks so it can be read better

* Update enum.md

* Update control-structures.md

Bump version + translating some of the code outputs in Greek

* Update modules.md

* Update comprehensions.md

Bump version

* Update strings.md

* Update plug.md

I believe elixir 1.4 introduced the change of applications -> extra_applications, where you don't need to add an app to start it, it is assumed from the dependencies. The original English lesson has the code chnaged but not the text around it. Should we remove this text, or let it with an asterisk as a backwards compatible option?

* Update eex.md

Bump version