Add benchmarking lesson
This adds a lesson on benchmarking, which I felt went well with the
lesson on debugging, so I added it below that lesson in the sidebar.
Since the existing function for benchmarking in Erlang (`timer.tc`) is
super minimal and not entirely human friendly, I decided to go with
Benchee since it's currently the best library for benchmarking in
Elixir.