Clarify frontpage language to note that right side entries are of the blog
Right now, if a new visitor comes it reads 'Recent Posts`juxtaposed with the book content of the left side once opening the index.
It is clear this is reference material there and to be consumed, but what is the right side and lower?
We do not translate blog entries at this time, we do concentrate on the content for obvious reasons.
This makes the boundary more clear.