Don't suggest that the concept of an interface will be familiar to Ruby developers
The "Specifications and types" page compares Elixir's
"specifications" to the concept of an "interface", suggesting
that this might be a helpful analogy if you're familiar with Java
or Ruby.

Ruby doesn't have a concept of interfaces in the language, so this
seems like a bit of an odd thing to say. (Although the generic,
abstract concept of an "interface", as in a "contract", is likely
to be familiar to many programmers.)

This removes this reference to Ruby, just referring to Java
instead.