Explain better the use of exceptions with a `:plug_status` field in the "Plug" guide
At the moment, the "Plug" page is a bit confusing in that it tells
you to add a `:plug_status` field to your custom exception when
it won't work until much later in the guide.

This adds the `:plug_status` field much later, after we've set up
the `Plug.ErrorHandler` plug, and explains how it works a bit
better.