Provide more useful exampels (#1849)
I think that using data which can not be changed is not clearly explain ideas of described rules. When we using a function for describing interpolation it is a good idea to use data that can be changed when that function was implemented. 
I offer a clearer explanation in the form of an improvement in the example used.