Add a small reference to the pipe operator (#2094)
* Add a small reference to the pipe operator 

One thing that really bothered me throughout learning is that the pipe operator wasn't mentioned during the first few chapters in the basics. Therefore I propose a change in order to make the lives of newcomers easier.

* Updated the commit based on a constructive comment

Based on the constructive commit to change this to what a newcomer would expect, I've removed the pipe operator reference and instead made `is_atom` look like a "regular function". This should clear out a lot of confusion.