Add capture operator usage to Enum.md
Adds instruction and examples on the use of the capture operator with the Enum module. This is a commonly used syntax and is not heavily referenced in ElixirSchool.

Many developers new to elixir have difficulty working with the capture operator and use ElixirSchool as their reference of choice. 

Thanks for any consideration!